+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add `org.jetbrains.intellij.deps.kotlinx` as a `kotlinx.coroutines` shadowed dependency group eligible for exclusion. 
+- Expand `TestFrameworkType` with additional frameworks including Debugger, ExternalSystem, PolySymbols, Qodana, UAST, VCS, and WebSymbols.
 
 ## [2.10.4] - 2025-11-02